Migrates broadcast_to to numpy_compat
